openqasm

Module ast

Structs

FileSpan

Represents a span of code in a file.

Program

Represents a whole program with defintions and statements. The definitions and declarations may cover more than one file.

Reg

A reference to (or definition of) a register or qubit.

Span

An object with an attached span. The span references where in the source code this object was derived from.

Symbol

An interned string constant.

Enums

Decl

A declaration of some kind.

Expr

A parameter expression.

Stmt

A statement that represents an action.