openqasm

Module translate

Structs

Linearize

High-level translation interface.

LinearizeError

An error that occured while linearizing a program. This error contains both the call stack leading up to this error, as well as the kind of error that occured. stack is layed out as a list of function names and call for those functions, from oldest to youngest (i.e the current stack frame is the end of stack).

Value

A value of type a + b * pi, with a, b rationals.

Enums

Binop

A binary operation encountered in an Expr.

LinearizeErrorKind

The type of error that occurred while linearizing.

Unop

An unary operation encountered in an Expr.

Traits

ExprVisitor

Companion trait to ProgramVisitor.

GateWriter

Output format from Linearize.

ProgramVisitor

Low-level translation interface for definitions and declarations.