openqasm::translate

Struct Linearize

pub struct Linearize<T> { /* private fields */ }

High-level translation interface.

This structure is used to turn a program into a linear list of gates. It uses a value that implements GateWriter to output primitive gates to some medium.

This can be used by constructing one with Linearize::new with your chosen GateWriter, and then using the ProgramVisitor impl to call .visit_program on your program. The depth argument to Linearize::new determines how many layers of definitions it will expand.

Note that it is assumed you have type-checked your program first. If you haven’t you may get garbage output / random panics. If you have type-checked it, you shouldn’t get any panics.

Example:

struct GatePrinter;

impl GateWriter for GatePrinter {
    type Error = std::convert::Infallible;
 
    fn initialize(&mut self, _: &[Symbol], _: &[Symbol]) -> Result<(), Self::Error> {
        Ok(())
    }

    fn write_cx(&mut self, copy: usize, xor: usize) -> Result<(), Self::Error> {
        println!("cx {copy} {xor}");
        Ok(())
    }

    fn write_u(&mut self, theta: Value, phi: Value, lambda: Value, reg: usize) -> Result<(), Self::Error> {
        println!("u({theta}, {phi}, {lambda}) {reg}");
        Ok(())
    }

    fn write_opaque(&mut self, name: &Symbol, _: &[Value], _: &[usize]) -> Result<(), Self::Error> {
        println!("opaque gate {}", name);
        Ok(())
    }

    fn write_barrier(&mut self, _: &[usize]) -> Result<(), Self::Error> {
        Ok(())
    }

    fn write_measure(&mut self, from: usize, to: usize) -> Result<(), Self::Error> {
        println!("measure {} -> {}", from, to);
        Ok(())
    }

    fn write_reset(&mut self, reg: usize) -> Result<(), Self::Error> {
        println!("reset {reg}");
        Ok(())
    }

    fn start_conditional(&mut self, reg: usize, count: usize, value: usize) -> Result<(), Self::Error> {
        println!("if ({reg}:{count} == {value}) {{");
        Ok(())
    }

    fn end_conditional(&mut self) -> Result<(), Self::Error> {
        println!("}}");
        Ok(())
    }
}

fn main() {
    let program = ...; // acquire a program from somewhere.
    program.type_check().unwrap(); // make sure to type check.
    let mut l = Linearize::new(GatePrinter, usize::MAX);
    l.visit_program(&program).unwrap();
}

Implementations

impl<T: GateWriter> Linearize<T>

pub fn new(writer: T, depth: usize) -> Linearize<T>