openqasm::translate

Trait ExprVisitor

pub trait ExprVisitor {
    type Output;

    // Required methods
    fn binop(
        &mut self,
        op: Binop,
        a: Self::Output,
        b: Self::Output,
    ) -> Self::Output;
    fn unop(&mut self, op: Unop, a: Self::Output) -> Self::Output;
    fn lookup(&mut self, var: &Symbol) -> Self::Output;
    fn pi(&mut self) -> Self::Output;
    fn int(&mut self, val: u64) -> Self::Output;
    fn real(&mut self, val: f32) -> Self::Output;

    // Provided methods
    fn visit_expr(&mut self, expr: &Span<Expr>) -> Self::Output { ... }
    fn visit_var(&mut self, var: &Symbol) -> Self::Output { ... }
    fn visit_binop(
        &mut self,
        op: Binop,
        a: &Span<Expr>,
        b: &Span<Expr>,
    ) -> Self::Output { ... }
    fn visit_unop(&mut self, op: Unop, a: &Span<Expr>) -> Self::Output { ... }
}

Companion trait to ProgramVisitor.

This visits an Expr piece by piece and uses the required callbacks to compute a return value. The type of the return value is given by ExprVisitor::Output.

The default visit_ methods just traverse the structure and call the callbacks where necessary, so if you override them, you must do the traversal yourself.

For example, if you wanted to compute the maximum integer value that appeared in an expression, you could do:

struct MaxFinder;

impl ExprVisitor for MaxFinder {
    type Output = u64;

    fn binop(&mut self, _: Binop, a: u64, b: u64) -> u64 { a.max(b) }
    fn unop(&mut self, _: Unop, a: u64) -> u64 { a }
    fn lookup(&mut self, _: &Symbol) -> u64 { 0 }
    fn pi(&mut self) -> u64 { 0 }
    fn int(&mut self, val: u64) -> u64 { val }
    fn real(&mut self, val: f32) -> u64 { 0 }
}

fn main() {
    let expr = ...; // acquire an expression from somewhere.
    println!("max = {}", MaxFinder.visit_expr(&expr));
}

Required Associated Types

type Output

Required Methods

fn binop(&mut self, op: Binop, a: Self::Output, b: Self::Output) -> Self::Output

fn unop(&mut self, op: Unop, a: Self::Output) -> Self::Output

fn lookup(&mut self, var: &Symbol) -> Self::Output

fn pi(&mut self) -> Self::Output

fn int(&mut self, val: u64) -> Self::Output

fn real(&mut self, val: f32) -> Self::Output

Provided Methods

fn visit_expr(&mut self, expr: &Span<Expr>) -> Self::Output

fn visit_var(&mut self, var: &Symbol) -> Self::Output

fn visit_binop( &mut self, op: Binop, a: &Span<Expr>, b: &Span<Expr>, ) -> Self::Output

fn visit_unop(&mut self, op: Unop, a: &Span<Expr>) -> Self::Output