openqasm::translate

Trait GateWriter

pub trait GateWriter: Sized {
    type Error: Error + 'static;

    // Required methods
    fn initialize(
        &mut self,
        qubits: &[Symbol],
        bits: &[Symbol],
    ) -> Result<(), Self::Error>;
    fn write_cx(&mut self, copy: usize, xor: usize) -> Result<(), Self::Error>;
    fn write_u(
        &mut self,
        theta: Value,
        phi: Value,
        lambda: Value,
        reg: usize,
    ) -> Result<(), Self::Error>;
    fn write_opaque(
        &mut self,
        name: &Symbol,
        params: &[Value],
        args: &[usize],
    ) -> Result<(), Self::Error>;
    fn write_barrier(&mut self, regs: &[usize]) -> Result<(), Self::Error>;
    fn write_measure(
        &mut self,
        from: usize,
        to: usize,
    ) -> Result<(), Self::Error>;
    fn write_reset(&mut self, reg: usize) -> Result<(), Self::Error>;
    fn start_conditional(
        &mut self,
        reg: usize,
        count: usize,
        val: u64,
    ) -> Result<(), Self::Error>;
    fn end_conditional(&mut self) -> Result<(), Self::Error>;
}

Output format from Linearize.

This trait is used by Linearize to actually output the linearized program. It contains various methods to output a primitive operation. Qubits and bits are referenced by consecutive integers starting from zero, and parameters are provided as Values.

The only non-obvious methods are initialize, which provides the names of qubits and bits to the backend, and is called exactly once, before any other function. start/end_conditional is called before and after any statements that are intended to be conditional on a classical value.

Required Associated Types

type Error: Error + 'static

Required Methods

fn initialize( &mut self, qubits: &[Symbol], bits: &[Symbol], ) -> Result<(), Self::Error>

fn write_cx(&mut self, copy: usize, xor: usize) -> Result<(), Self::Error>

fn write_u( &mut self, theta: Value, phi: Value, lambda: Value, reg: usize, ) -> Result<(), Self::Error>

fn write_opaque( &mut self, name: &Symbol, params: &[Value], args: &[usize], ) -> Result<(), Self::Error>

fn write_barrier(&mut self, regs: &[usize]) -> Result<(), Self::Error>

fn write_measure(&mut self, from: usize, to: usize) -> Result<(), Self::Error>

fn write_reset(&mut self, reg: usize) -> Result<(), Self::Error>

fn start_conditional( &mut self, reg: usize, count: usize, val: u64, ) -> Result<(), Self::Error>

fn end_conditional(&mut self) -> Result<(), Self::Error>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.