openqasm::typing

Enum TypeError

pub enum TypeError {

    RedefinedGate {
        name: Span<Symbol>,
        decl: FileSpan,
        prev: FileSpan,
    },
    RedefinedArgument {
        name: Span<Symbol>,
        decl: FileSpan,
        prev: FileSpan,
    },
    RedefinedParameter {
        name: Span<Symbol>,
        decl: FileSpan,
        prev: FileSpan,
    },
    RecursiveDefinition {
        cycle: Vec<(FileSpan, FileSpan)>,
    },
    ZeroSizeRegister {
        reg: FileSpan,
        decl: FileSpan,
    },
    RedefinedRegister {
        reg: Span<Reg>,
        decl: FileSpan,
        prev: FileSpan,
    },
    UndefinedParameter {
        name: Symbol,
        stmt: FileSpan,
        span: FileSpan,
    },
    InvalidComparisonSize {
        reg: FileSpan,
        value: FileSpan,
        stmt: FileSpan,
        reg_size: u64,
        value_size: u64,
    },
    WrongArgumentArity {
        stmt: FileSpan,
        def: FileSpan,
        name: Symbol,
        arity: usize,
        actual: usize,
    },
    WrongParameterArity {
        stmt: FileSpan,
        def: FileSpan,
        name: Symbol,
        arity: usize,
        actual: usize,
    },
    UndefinedGate {
        name: Span<Symbol>,
        stmt: FileSpan,
    },
    WrongOperandSize {
        stmt: FileSpan,
        span: FileSpan,
        size: u64,
        match_span: FileSpan,
        match_size: u64,
    },
    InvalidRegisterIndex {
        name: Symbol,
        index: u64,
        size: u64,
        reg: FileSpan,
        def: FileSpan,
        stmt: FileSpan,
    },
    WrongRegisterType {
        name: Symbol,
        def: FileSpan,
        stmt: FileSpan,
        reg: FileSpan,
        classical: bool,
    },
    UndefinedRegister {
        name: Symbol,
        reg: FileSpan,
        stmt: FileSpan,
    },

}

An error produced during type-checking.

This includes several main categories:

Redefined...: two objects have the same name,

Wrong...: the arity/type/size doesnโt match what is required,

Undefined...: the thing referred to doesnโt exist,

and a few other miscellaneous errors.

In general, name refers to the name of the thing that is redefined/undefined/wrong, while def/decl refers to the place it is defined, stmt refers to the statement where this error occured, and prev is where something was previously defined or referenced. reg/span refers to the object in question.

Documentation is given for fields that are not listed here.

Variants

RedefinedGate

This gate has already been defined.

Fields

name: Span<Symbol>

decl: FileSpan

prev: FileSpan

RedefinedArgument

This argument name has already been used.

Fields

name: Span<Symbol>

decl: FileSpan

prev: FileSpan

RedefinedParameter

This parameter name has already been used.

Fields

name: Span<Symbol>

decl: FileSpan

prev: FileSpan

RecursiveDefinition

This sequence of definitions is recursive.

Fields

cycle: Vec<(FileSpan, FileSpan)>

These definitions form a cycle. The first component refers to the definition, the second component to where it calls the next definition in the chain.

ZeroSizeRegister

A register has been declared with size zero.

Fields

reg: FileSpan

decl: FileSpan

RedefinedRegister

This register has already been declared.

Fields

reg: Span<Reg>

decl: FileSpan

prev: FileSpan

UndefinedParameter

This parameter is undefined.

Fields

name: Symbol

stmt: FileSpan

span: FileSpan

InvalidComparisonSize

This comparison is out of range.

Fields

reg: FileSpan

The classical register being compared.

value: FileSpan

The constant value it is compared to.

stmt: FileSpan

reg_size: u64

The size of the register.

value_size: u64

The size of the constant value.

WrongArgumentArity

This statement has the wrong number of arguments.

Fields

stmt: FileSpan

def: FileSpan

name: Symbol

arity: usize

The correct number of arguments.

actual: usize

The actual number of arguments.

WrongParameterArity

This statement has the wrong number of parameters.

Fields

stmt: FileSpan

def: FileSpan

name: Symbol

arity: usize

The correct number of parameters.

actual: usize

The actual number of parameters.

UndefinedGate

This gate is undefined.

Fields

name: Span<Symbol>

stmt: FileSpan

WrongOperandSize

This operandโs size doesnโt match the others in this statement.

Fields

stmt: FileSpan

span: FileSpan

Reference to this operand.

size: u64

The size of this operand.

match_span: FileSpan

Reference to the operand it needs to match.

match_size: u64

The size of the operand it should match.

InvalidRegisterIndex

This index is invalid for this register.

Fields

name: Symbol

index: u64

The index being accessed.

size: u64

The size of the register.

reg: FileSpan

def: FileSpan

stmt: FileSpan

WrongRegisterType

This operand has the wrong type.

Fields

name: Symbol

def: FileSpan

stmt: FileSpan

reg: FileSpan

classical: bool

Whether or not the operand is supposed to be classical.

UndefinedRegister

This register is undefined.

Fields

name: Symbol

reg: FileSpan

stmt: FileSpan